Fix session creation failing unconditionally on Postgres
sessions.rotation_counter is declared INTEGER (32-bit) in the migration,
but SessionRow read it as i64. SQLite's untyped storage let this pass
every test in the suite, since every test runs against an in-memory
SQLite database; sqlx's Postgres decoder, by contrast, rejects a struct
field wider than the column it reads from and rejects the row outright.

The practical effect: create_session failed on every call against a real
Postgres database, which meant issue_session (and therefore
complete_login) failed on every login that actually needed to establish
a session — including, critically, every MFA challenge/enrollment
completion, since those only issue a session at the very end of the flow.
This is what was actually behind the "something went wrong" error
reported on the live deployment; retrying only ever hit the same failure
underneath, no matter how the retry itself was handled.

Fixed by matching the Rust field width to the actual column width. Added
an opt-in Postgres-backed regression test (mirroring the existing
pgvector_search.rs convention) so this class of Postgres-only decode
mismatch — invisible to the SQLite-backed suite that runs in CI — gets
caught going forward.

+++ server/src/db/session.rs
@@ -34,7 +34,13 @@ pub struct SessionRow {
     pub revoked_at: Option<String>,
     pub user_agent: Option<String>,
     pub ip_address: Option<String>,
-    pub rotation_counter: i64,
+    // Matches the Postgres column's actual width (`INTEGER`, 32-bit) —
+    // sqlx's Postgres decoder rejects a struct field wider than the
+    // column it reads from (unlike SQLite, which is untyped and would
+    // have accepted `i64` here without complaint, hiding this on every
+    // SQLite-backed test run). A session realistically never rotates
+    // anywhere near i32::MAX times, so this is not a real capacity limit.
+    pub rotation_counter: i32,
     pub created_by: Option<String>,
 }
